requirements.txt
pandas==2.2.2
requests>=2.32.3
python-dateutil>=2.9.0
